Fix a quick issue with market transactions
I recently added the tracking of 'profits' per sale and while that worked it was wrong, it didn't actually work and it led to a small improvement.

I was always overwriting the entire market_transaction record and I didn't take that into consideration.  So after processing the record it gets the profit but next data refresh wipes it out because that incoming record doesn't have this data.  So, on the receiving end we're now checking to see if that transaction id exists before saving.

On the processing side we're no saving the value as a profit per item and not profit for the transaction as a whole and that's how I say it was wrong.  Just made no sense to have a record with a quantity and that quantity is not applied to a numeric value in that record.