Fix an issue with product setup
The product setup procedure was off.  The first step - retrieving the bp details was incorrect and wasn't populating the blueprint type id.  The second portion was in the setup part and it was setting the makeType on the product to the product type id so whenever one was built one was removed...